Add better defaults to Assumptions as scenarios
Users will be more likely to know the names of the
assumptions they are using, so we will accept
those as the inputs to the Scneario creation, and
as such, the other fields of the Assumption schema
need to have good defaults. The schema will fail
if the user doesn't add all data when adding a new
assumption.