Refactor forms to build more from python objects
Each page gets its own jinja template, which is reposible for placing
content. For example, all edit pages contain a list of existing objects,
and one new object form. The content of these forms is determined by the
objects passed to the template. Those objects implement a `render`
method, which is called. This allows for the logic of what to render to
be written in python, and not in jinja. This also allows for a type of
dependency injection.

This will come in really handy when we attempt to implement Scenarios,
since the format of that form will reqire a dynamic number of dropdown
menus. Implemnting that logic in jinja would have been error prone.